backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.0
openai>=1.0.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0
requests>=2.31.0
